requirements.txt
aerich==0.6.3
aiofiles==0.8.0
aiomysql==0.1.0
aiosqlite==0.17.0
click==8.1.2
colorama==0.4.4
dictdiffer==0.9.0
httptools==0.4.0
iso8601==1.0.2
multidict==5.2.0
pydantic==1.9.0
PyMySQL==1.0.2
pypika-tortoise==0.1.4
pytz==2022.1
sanic==21.12.1
sanic-routing==0.7.2
tomlkit==0.10.1
tortoise-orm==0.19.0
typing_extensions==4.1.1
websockets==10.2
